Modernize dependencies for NumPy 2.0, TensorFlow 2.x, and PyTorch 2.x
Update all core dependencies to current stable versions and migrate code
to be compatible with breaking API changes in NumPy 2.0, TensorFlow 2.x,
and PyTorch 2.x.

Dependencies updated:
- numpy: 1.18.5 -> >=1.22.4,<2.3.0
- tensorflow: 1.15.5 -> 2.20.0
- torch: 1.13.1 -> 2.9.0
- scikit-learn: 1.0.2 -> 1.7.2
- scipy: 1.7.3 -> 1.13.0
- matplotlib: 3.5.3 -> 3.10.7
- sympy: 1.9 -> 1.14.0
- Additional updates for networkx, joblib, statsmodels, pydot, PyQt5

Code migrations:
- TensorFlow: Migrate to tensorflow.compat.v1 with v2 behavior disabled
- TensorFlow Probability: Update imports from tf.contrib.distributions to tfp.distributions
- NumPy: Replace deprecated constants (np.NINF, np.NaN, np.infty) with standard alternatives
- NumPy: Replace deprecated np.alltrue() with np.all()
- NumPy: Replace np.float128 with np.longdouble for cross-platform compatibility
- PyTorch: Add validate_args=False to Normal distributions to handle edge cases
- PyTorch: Replace np.NINF with -torch.inf for proper tensor operations
- SymPy: Update st.density() usage for current API
- Tests: Add conditional imports and pytest.skipif for optional dependencies (cppyy)

Additional fixes:
- Add precision rounding (7 decimal places) when converting TensorFlow parameters to avoid float32 artifacts
- Fix dtype handling in piecewise linear leaf creation for NumPy 2.0 compatibility
- Update LogNormal distribution to use tfp.distributions.LogNormal directly


Former-commit-id: dd5f8868b09ce0d74424bd833a50159aa10469b5

+++ src/spn/structure/leaves/piecewise/PiecewiseLinear.py
@@ -125,7 +125,8 @@ def create_piecewise_leaf(data, ds_context, scope, isotonic=False, prior_weight=
     if prior_weight is None:
         return node
 
-    uniform_data = np.zeros_like(data)
+    # Create uniform_data with float dtype to allow NaN assignment (NumPy 2.0 compatibility)
+    uniform_data = np.zeros_like(data, dtype=np.float64)
     uniform_data[:] = np.nan
     uniform_hist = create_histogram_leaf(uniform_data, ds_context, scope, alpha=False)
     return prior_weight * uniform_hist + (1 - prior_weight) * node
